fix: change compute_dtype from float16 to float32 for compatibility

+++ 05_test.py
@@ -2,7 +2,7 @@ from dia.model import Dia
 import torch
 
 dia = Dia.from_pretrained("nari-labs/Dia-1.6B", device="cuda",
-                          compute_dtype="float16", load_dac=True)
+                          compute_dtype="float32", load_dac=True)
 dia.model.load_state_dict(torch.load("dia_crying_finetuned.pt"), strict=False)
 
 txt = "[S1] I’m sorry… (crying) [S2] Please calm down."
